v3.1.0 - Version update.
The new version has been set in libdue, so...
 - Tarfile generation will get it automatically.
 - Docs have been rebuilt.
 - Debian changelog has been updated

The version bump is warranted by a number of fixes for Podman
compatibility, with the big one being that the user's identity
wasn't created in the container, so their home directory was
not writeable.

Additionally, fixes were added for RedHat container compatibility,
in that the version of 'adduser' in Fedora 33 does not support the
same arguments as the version of 'adduser' in Debian.
I suspect these sorts of compatibility issues will come up again,
so this may provide a useful template for handling them.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>