onie/post-install-config.sh.template - replace wget with curl
Closes #37 - Creating default Debian 8 build environment fails

This addresses an issue where the 'hub' open source pacakge is
pulled in from GitHub. It is used for processing pull requests,
and as such isn't tied directly to end user ONIE builds, but
is more of an adminstrative tool.

Wget was quickly iterating through redirects, and would quit after
hitting its maximum of 20 when downloading the hub tar.gz file.
Insofar as I know, this is new behavior, so I'm assuming something
changed on the GitHub end. Regardless, it should be handled
without issue.

The solution was to use curl, which doesn't seem to have this problem.

Testing:
--------
Built Debian 8->11 images without issue.
Built the only ONIE target that uses Debian 8 with a Debian 8
using the 2021.08 branch of ONIE.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>