due - Be very explicit about local sourcing
If DUE is not run from it's system installed location,
AND ./libdue is local, then source that libdue, and print
a notice that it is being used.

This fixes a bug in the previous logic where supplying an absolute
path to a local copy of due would pull in libdue from the directory
it was being invoked from.

This will still provide the ability to use a locally checked out
copy of DUE, rather than the one installed on the system, and it will
be very clear about which files are being used.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>