libdue - Allow use of qemu-user-static from template's filesystem/usr/local/bin
This adds some sanity checking to building qemu-user-static in to the
image, and allows for a copy of qemu-user-static to be pulled in from
the template's filesystem/usr/local/bin directory.  This is particularly
handy if one wants to use a newer version of qemu than is installed on
the host system.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>