libdue - handle empty DUEMountHostDirsUp
DUEMountHostDirsUP is a LABEL used in Dockerfile.config by
DUE to change the mount point of the current working directory.

Debian templates use this to mount the directory ABOVE the one
where the build is occurring, as Debian packages are copied there,
and if the container doesn't mount that, they vanish with the container,
rather than staying on the host system as expected.

This adds logic to handle an empty value being returned, as not all
containers will require this, and it being empty is legitimate.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>