libdue - --login command now runs container-create-user.sh
Previously --login just ran a bash login shell. Now it runs
container-create-user.sh the same way --run does.

If the user that started the container logs in again, they
should be in their home directory with all configuration applied,
keeping the behavior consistent.

If a different user logs in, they will get an account created, and
will be able to sudo, but will not have their home directory mounted.
Since the container is already running, that's not going to happen.
They will get a /home/$USER directory created in the container, but
they shouldn't get too comfortable as that will vanish when the
container does.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>