libdue - clean up run and login command output formatting.
This makes a few updates to DUE's echoing of the command that
it ends up running to start a new container, or log in to an existing one.

Run format was updated to not have spaces after the end of echoed '\' so that
the command can be cut and pasted elsewhere to run the same way.

Login now uses box lines drawn with --- to indicate that the context is DUE's
execution, and not the container ( run has done this since day 1)
Login also echos a formatted command string that can be cut and pasted, to
make it consistent with run.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>