due.conf - Allow default run time arguments on a per container type basis.
The ability to supply default arguments when running a particular
type of container is extremely useful. Whether that is used to
automatically mount a host directory in the container, or run
--privileged, or pass along the address of a server, it is one less
thing the end user has to keep track of.

Implementation to this point was by editing libdue and modifying
 fxnSetContainerSpecificArgs to switch off the DUEImageType
Docker LABEL (set in the Dockerfile.config file under the templates
directory) and append arguments as needed.

However, requiring edits to a file under /usr/lib is not really ideal,
so now libdue sources the users (or the system's) due.conf file to fill
out the actions to be taken on particular containers.

This should allow for a system-wide deployment of behaviors (/etc/due.due.conf)
or user customization (~/.config/due/due.conf), and persist them through
DUE upgrades.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>