due.conf config file clean up and handling improvements.
As the conf file is now being sourced, this removes the routine
that used to parse it, and sets a the user's home directory to ~/
as a default until overridden.

DUE now sources the system's /etc/due/due.conf file first, and
it _may_ source user's ~/.config/due/due.conf files if the value
of DUE_ALLOW_USER_VERSION in /etc/due/due.conf is set TRUE.

By default it is set FALSE to allow the sysadmin to configure
settings and default arguments system-wide that cannot be changed
by users.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>