Update --run --debug settings.
This sets --run and --login to have the user enter the container as
root, and run /bin/sh by default.
Since this invocation is expected to be used with broken or non-DUE
containers, few assumptions should be made about whatever software
may have actually been installed.

One key change is the use of the DEBUG_LOGIN variable (rather than DO_DEBUG)
to indicate the reduced parameters in the login. DO_DEBUG will
invoke a set -x, or pass that along to other scripts, which is not
what this case needs.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>