SUSE cleanup - OpenSUSE and SLES share post-install-config.sh.template
As the post install for OpenSUSE or SLES is mostly identical, the
post install script has logic to change a few intalled packages based
on the target OS, which allows it to be shared by both targets.

This also provides an example of the directory inheritance model
DUE uses to reduce the number of redundant (and mostly redundant)
files one might need. By placing post-install-config.sh at the same
level as the sub-type directory it gets pulled in by all container
types below that directory.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>