pre-install-config.sh.template - Shellcheck cleanup and code unification.
This includes:
 - Shellcheck suggested changes.
 - Reduction of redundant code by parsing Red Hat/SUSE package handling
    as one function since that is all RPM manipulation.
 - Added processing of pre-install-local files for Red Hat/SUSE.

While actions in this script could be implemented as separate specific
scripts under the appropriate template directories (as they would replace
this script), the script doesn't do much, so it is simpler to keep it
as an operating system aware file.

Besides, as a shell script, it's main purpose is to make sure bash is there
for the other scripts to work.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>