Fedora post-install - use dnf post install and whitespace cleanup.
This fixes some whitespace cleanup and uses dnf in post install of
local packages for consistency (as it is used for initial install),
and because it and yum will consider all of the RPMs passed when
resolving dependencies.

If building for older releases, it may be a good idea to have logic
switch between the older 'yum' and 'dnf'.  That can be addressed when
a use case comes up.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>