Makefile - expand OS/Docker support for make install
'make install' allows a user to install DUE without building a
package (especially useful as DUE doesn't build as an RPM...yet)

This expands the Makefile's awareness of package installers to
use (apt/dnf/zypper) and whether to install Podman or Docker.io
by default.

Docker.io is shipped with a distribution and does not come
directly from Docker.

Podman is preferred by Red Hat, and so their systems are less likely
to have docker.io

So when installing.

If Docker or Podman is already installed, leave it.
If this is a Red Hat system (the /etc/os ID_LIKE is = fedora) default
to Podman
Otherwise default to using docker.io.

And if the user is on Windows Subsystem for Linux, that'll be Debian,
but because of limited daemon support, the user should pre install
Podman, and then run this because the Makefile has no way of knowing
this is the case.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>