Makefile - fix OS detection bug when using RHEL
The /etc/os-release in RHEL had quotes around the ID_LIKE value,
which is no big deal in Bash, but crippling in a Makefile that
does not expect them to be there - so that gets filtered out now.

Also consolidated the OS specific install logic in to a per os type
grouping at the top to make it clearer what is getting set, and
easier to change it later on.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>