libdue - Retain user identity with Podman
This had been fixed at some point way back, so for clarity:

Podman will want to remap user space to create a user account that
is not associated with any on the host system. As the whole point
of DUE is to retain one's identity in a container, this gets worked
around.

So:
 1 --userns keep-id is used to keep the name space remapping
                    from happening.
 2 --user 0:0 is used to log the user in to the container as root,
              so that container-create-user.sh has the ability
              to create user accounts, and will create an account
              for the user who launched the container.

Testing:
 Was able to access ~/ on my host system
 Was able to write a file in to ~/
 On the host, that file had my host username and host group, and
 was readable from the host.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>