libdue - identifying containers in Podman versions less than 4
If Podman is used, this change will detect major versions less than
4, and will launch containers using their name:tag rather than their
ID.
I hit this bug in Podman 3 in Windows Subsystem for Linux. I'm
not exactly sure where the dividing line is for versions, as I had just
the two to test with, so this may be refined further.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>