due,libdue - clean up sourcing messages.
Reworked the sourcing such that no sourcing messages are
printed if DUE is using the system default files located
in /lib, /bin and /etc/due respectively.

If DUE is being run locally with a ./due, then a message will
print at the start of the run that notes that.
If /etc/due/due.conf has the DUE_ALLOW_USER_CONFIG variable
set to "TRUE", then DUE will print the path to the invoking
user's ~/.config/due/due.conf.

Testing:
--------
 Run installed due with 'due'
 Run local checkout with ./due
 Edit /etc/due/due/conf to set:
   DUE_ALLOW_USER_CONFIG="TRUE" , and have a ~/.config/due/due.conf.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>