Makefile - clean up basename error, make OS detect clearer.
This sends the error thrown by basename to /dev/null if it
is passed an empty value for the path to Podman. As the desired
end result is to have an empty variable for podman if it is not
present, then this is fine.

Also cleans up the message that indicates which host OS was
detected. Printing the value from /etc/os-relase is rather
cryptic if one doesn't know the context, so displaying it as
the detected OS will be more comprehensible to the average user.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>