RedHat duebuild - Start cleanup
This rolls in some save-it-so-they're-not-lost changes for kernel
patching as well as format cleanup and fixing of the --default
argument behavior.

It's way more than I'd like to see in a commit, but as this is
still in development I'm not seeing a return on the effort of
going through all the changes. Plus by locking that chaos down
now, subsequent changes will be more easily identified.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>