For debug, --run with --debug should override container's Entrypoint
This adds logic such that a container that was not created with DUE
can be run by DUE for debug purposes which requires ignoring the
contianer's Entrypoint, and replacing it with an interactive shell
so the invoking developer can take a look around.

The shell defaults to /bin/sh, which should be available in most
instances.  If the user wants to put in the extra effort to upgrade,

./due --run --debug --login-shell /bin/bash

Can be used to invoke a bash shell, if it is already installed in
the container.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>