Makefile - add conditional to be inactive during package build.
This adds a check for DUE_BEING_PACKAGED_BUILD being set to
allow a top level Makefile to co-exist with a ./debian
directory during development/debug.

Background:
-----------
During development, I frequently have a ./debian directory
to be able to build a test package along with a Makefile for
brute force installation. If package build finds a Makefile,
it will (quite reasonably) try to run it, but this throws errors
as the DUE Makefile is there to install DUE, not build.

So this change has the top level Makefile do nothing if
DUE_BEING_PACKAGED_BUILD is set to TRUE, and allows me to build
a test .deb without changing git branches or moving files around,
which always leads to more work as something gets messed up.

More details:
-------------

A ./debian/rules file can set the variable with
 export DUE_BEING_PACKAGED_BUILD = TRUE
OR
 It can be set on the shell command line with
 DUE_BEING_PACKAGED_BUILD="TRUE" make

If set, the Makefile does nothing, prints a message to
this effect and packaging continues.

Otherwise, all Makefile targets are valid.

The catch here is that the DUE Makefile mostly exists to
install DUE in the 'checked out from source' case, so no package
manager is involved, and invoking the install commands during packaging
throws errors and dies. So the solution is to just not do this.

Release details:
----------------
The debian/master branch used for upstream release has not, does not and
will not have a top level Makefile to collide with it's debian/rules
file, which is why this is has not been an issue for previous upstream releases,
Thus the ./debian/rules file does not have to export any values ( and, indeed, this is
good, because if the Makefile is present, when it should not be,
we _want_ things to break.)

Signed-off-by: Alex Doyle <adoyle@nvidia.com>