Add --entrypoint override option
This allows the user to override the image's ENTRYPOINT - the command
run by default on container start up - with a shell, or, if a
shell is not supplied, DUE will try to run container-create-user.sh.

This is useful for using DUE to wrap containers that already have
a non-shell ENTRYPOINT defined, to make run time debugging easier, or
just getting a shell in a container that has a non-shell ENTRYPOINT.

As the specifics of the image being run can vary, this is not guaranteed
to work for all cases, but the followning scenarios have been tested.

Testing:
---------
Non-DUE contianer:
 due --run --entrypoint /bin/bash
  - ran bash as root

Container with ENTRYPOINT used as a start point for a DUE build,
so that container-create-user.sh was in the image
 due --run --entrypoint
  - ran container-create-user.sh and created the invoking account
    in the contianer.

Signed-off-by:
Alex Doyle <adoyle@nvidia.com>