Debian package post-install-config.template - Support python3-six in Sid.
Newer Debian releases do not have python-six, but the Python 3 version
of that: python3-six, which caused a post install error creating Sid based
images when python-six could not be found.

As the post-install-config.sh script that is genereated from the
post-install-config.template can be running on Debian releases as far
back as 8 Jessie, it has to be aware of this.

So the solution is to try and install classic Python packages, and
if that fails, fail silently and try for the newer Python 3 versions
to provide the expected level of functionality.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>