debian/ copyright and due.lintian-overrides updates
This updates the copyright and overrides file for Lintian checks
to test changes for the upcoming 4.0 upstreaming.

Note that the due.lintian-overrides file works fine in Debian 11,
but fails to be parsed correctly in Debian Sid. Lintian is parsing
the overrides file but has a series of mismatched overrides that
prevent it from taking effect.

Also note that the control file has been left with debhelper (>=10),
rather than the latest (>=13) required for upstreaming so that this
will build on older releases.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>

It's unknown at this time what's causing this.