Makefile - Do not recreate orig.tar on every build.
This changes the orig.tar target to NOT rebuild the 'upstream tarball'
orig.tar file used for packaging .deb and .rpm files for release.

Otherwise it would be touched between .deb and .rpm builds, so this
assures that .rpm and .deb releases have exactly the same code
(time stamps and all).

Creating a clean tar file is left as an exercise for the developer
to manually delete as they see fit, since the code trying to be clever
about this is likely to make it worse.

Also cleaned up some display and informational formatting in the output.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>