libdue - do not blame Docker fails on Podman systems.
This puts a conditional around an error that happens if Podman
errors out trying to list images. With Docker, it usually means
that the daemon isn't running, but since Podman doesn't have a daemon,
it really doesn't apply.

Found this testing rpm package install in a Fedora 36 container and
trying 'due --run' just to see what would happen.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>