due.lintian-overrides - update path formats for latest Lintian.
The latest Lintian in Sid as of release preparation is 2.115.3.

This version expects file paths used in Lintian overrides to
be enclosed in square braces [<path>], whereas the previous version
used in the 3.0.0 release of DUE did not.

This change adds square braces to the paths used to specify
Lintian overrides to eliminate 'mismatched tag' errors, and have
the overrides apply correctly.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>