due.spec - fix due.conf copy
Fix a cut and paste error to put etc/due/due.conf in
/etc/due/due.conf
on the target system,
Add /etc/due/due.conf to %files to fix 'unpackaged file' error.

Testing
-------
Tested on a fresh Fedora 37 install, as re-using a Fedora
VM let this problem slip by in the first place.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>