Makefile - Automate packaging build cleanup.
This adds error handling to the rpm-package and debian-package
targets, which check out the rpm/master and debian/master branches,
then do the packaging build, then return to the master branch.

When the build fails the build area is left on the rpm or debian master
branch, which may not be immediately obvious to the user, and it
has files scattered around it which make switching back to master,
and remembering to 'stash pop' any changes, a bit much to ask.

So, for either of those targets, the build will always clean up
after itself as I've added a ' ; true' to keep even failed builds
from exiting the make.

To be able to actually flag errors, the Makefile will delete any
existing .deb or .rpm files that would be produced before the build
starts. (This happens a few steps in to the build so that conditions
that would prevent the build from working (wrong branch, missing tarfile,
etc ) have to be satisfied before the old build gets deleted)
If the deb/rpm does not exist at the end of the build, it's a fail,
and a big failure message will be printed.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>