due.conf - Add sample defaults for ONIE build
This adds a commented out example of the settings I'd used producing
ONIE releases for ONIE containers. The two options are:

--privileged - needed to loopback mount filesystems, needed when doing
  Secure boot builds that set up a virtual USB drive.

--volume - mount the host's cache of ONIE downloaded packages under
  /var/cache/onie/downloads (only use it if it is there.) This
  can speed up builds by reducing the need to download packages off
  the internet.

The example is commented out as not all users will need it, and
not all systems will have a cache, so it is left to the developer to
know what the right thing to do is.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>