Makefile - fix build success logic.
This moves the check for a resulting .deb or rpm in to the second
pass of the Makefile, where it will execute AFTER build has been
attempted.

This does explain the failures on the first pass of the build.
If the .deb or .rpm was left over from a previous build, it would
be detected before build start, and everything would be fine.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>