libdue Export image name and tag when exporting an image
This modifies the docker|podman save command to use a name:tag
rather than an image Id, which preserves the name and tag of the
container in the exported image so that it can be used on import.

Note that images exported from Podman will import to Docker with
a 'localhost/' in front of the image name.
This gets in to the 'short names' of Docker images not presenting
the full URL of where the image came from.
Docker assumes 'Docker.io/' and removes this.
Podman defaults to 'localhost/'
...which is why images exported from Docker to Podman get
'localhost/' prepended to the image name, and since images
exported from Podman to Docker don't start with 'Docker.io',
the 'localhost/' shows up on Docker import.

As there doesn't seem to be a way to get around this without
'being clever' and introducing more corner cases to cover
(Ex: tag the image, then export it, then delete the tag...unless
a tag already exists...) I'll leave it at "Functions As Designed"

Testing:
--------
Export Podman generated image, load on Docker system.
Export Docker generated image, load on Podman system.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>