Version 4.1.0 Add image-patch directory, and --image-patch create option
User Malte Janduda (https://github.com/MalteJ) discovered that the latest
Debian 9 stretch images have an apt/sources.list file that no longer references
Stretch packages correctly since they nave been moved to archive.debian.org.
Additionally, stretch-updates has become stretch-proposed-updates.

(This article was useful in seein the to-archive conversion:
  https://www.geekersdigest.com/apt-etc-apt-sources-list-for-every-debian-release/ )

Malte submitted a pull request ( https://github.com/CumulusNetworks/DUE/pull/43 )
that would have addressed the ONIE build, which makes sense because DUE
(up to this point) has assumed that the base images being pulled could
perform a package update, or were otherwise totally functional.

However when the Debian 9 image went stale, it also broke any template that
uses Debian 9, so both the ONIE and debian-package templates are broken,
for Debian 9, stuck trying to access package repositories that are no longer there.

And this problem is only going to get worse as other images age out, or
repositories go offline, so a more general solution is required to allow
the end user to modify the starting state of the image. Package repository
access is the only problem case right now, but there may be others.

To solve this I'm adding an image-patch directory where updated files
for images can be stored, and will be copied in to an image build if
the --image-patch < path to image-patch directory>/<image-name>/<tag>/
argument is passed during creation.

Those files will get copied in to the build area, and will end up under
/due-configuration in the container.

The pre-install-config.sh script can then apply the changes before
anything template specific executes.

For example, the pre-install-config.sh template that generates the first
script that DUE runs in the container during configuration starts by
usin apt to install bash, sudo, and other programs DUE scripts expect.
Then it applies the template files which may do things like
overwrite sources.list with a copy that points to a developer local repository.

With this change, the pre-install-config.sh  now checks for
an etc/apt/patch-sources.list file, and, if present,
overwrites the container's /etc/apt/sources.list before trying to
apt update, and before the template files are installed.

The patch file starts with patch- both for clarity, and so it can
co exist with a  sources.list file the template may supply
when it's files are copied in to the image creation directory.

Code changes include:
 - Adding image-patches directory with updated sources.list for Debian 9
 - Adding --image-patch argument
 - Updating due and libdue to handle the argument and directory
 - Updating makefile to install image-patches directory
 - Updated pre-install-config.sh.template to deploy the sources.list
 - Updated documenation and man page
 - Rolled in Maltej's update to stretch-backports.list
 - Updated version to 4.1.0 to distinguish from the last release

Signed-off-by: Alex Doyle <adoyle@nvidia.com>