Update debian-package and example to mention Debian 12 Bookworm
This updates the README.md files for the debian-package and
example templates to reference Debian 12 rather than Debian 11
in the example strings they provide when:
 due --create --help
is run.

Nothing is wrong with Debian 11 - this just demonstrates that
the latest release is supported and those build configurations
work...and we don't need examples for every possible entry.

Testing: ran the following, logged in, and built with debian-package
---------
./due --create \
      --platform linux/amd64  \
      --name package-debian-12 \
      --prompt PKGD12 \
      --tag pkg-debian-12-amd64 \
      --use-template debian-package \
      --from debian:12 \
      --description "Package Build for Debian 12"

./due --create \
      --platform linux/amd64 \
      --name example-debian-12 \
      --prompt ExD12 \
      --tag example-debian-12 \
      --use-template example \
      --from debian:12 \
      --description "Debian 12 example"

Signed-off-by: Alex Doyle <adoyle@nvidia.com>