Makefile - updte help message for debian-test branch build.
The debian-test branch is used to test Debian packaging before
release to reduce the number of merges from master to the
debian/master branch.

Basically it's scratch space to figure out what one clean merge
for promotion will look like.

This change is a reminder to set the Makefile variable which
selects the branch to use when testing.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>