duebuild - handle auto generation of sources repositories for Deb822
So with Debian 12, the format of files specifying source repositories
changed to have a .sources at the end, and internal formatting is a
bit different.

The --download-src option will try to auto-generate corresponding
source repository entries if they don't exist.
This is common with most released containers as they don't default
to being build environments (hence, DUE...).

This change looks for a *.sources file under /etc/apt, and if it
finds one, it creates a copy with Type: deb changed to Type: deb-src,
and puts the copy under /etc/apt/sources.list.d, and runs apt update
to make the contents accessible.

Note that this is not intended to be completely robust at this point.
In most cases finding a single file to convert should work, and if
it isn't enough, the resulting configuration should be a good starting
point for debug, as it has examples of what the user should do.

Testing:
./duebuild --download-src ncdu

Worked in a Debian 12 and an Ubuntu 18 container.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>