libdue - Remove merge directory after a menu create
As menu creation ( due --create --menu ) invokes a pre-existing
configuration pulled from the template README files, the implication
is that the user is not actively debugging the creation of an image,
and just wants it created.

So there is no need to keep the merge directory around, and having it
there can trip up subsequent creation invocations.

The use case was a network mounted home directory that was used to
install DUE on different reserved lab systems, so that the invocation
directory remained the same while the install system changed, requiring
the recreation of the same image for each new reservation, and the
last merge directory would have to be deleted before the new image could
be created for the new system.

Testing:
--------
Created an example container, and directory was deleted
Created a root owned merge directory before trying container creation,
and, as expected, got errors about being unable to write to the directory.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>