Add option to merge host's /etc/group file with the container's
This was a feature I'd worked on that turned out to not be needed.
However it was enough work to figure out that I'm adding it in case
there is a use case where it is necessary to merge a custom
/etc/group file with that of the container's
(perhaps adding addtional groups the user is in?)

While the merge seems to work and not hurt anything, I don't have a
test case for it at the moment, so I'm leaving it undocumented, and,
since the group file under filesystem/etc gets merged in, rather
than copied directly, it will be named due-group to prevent accidental
invocation of the feature.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>