Troubleshooting.md mention everything-in-usr.
Mention that install failures that mention 'usrmerge' are probably
due to an older Docker image colliding with newer code that expects
some top level directories to be links to /usr.

A web search for 'everything-in-usr' will provide more details on
the reasons for the change, which are outside of DUE's scope as they
need to be supported.

The triggering case was DUE using a 2 year old cached Sid (Debian
unstable) image that did not play well with current software
that expected a different file system layout.

Deleting the Sid image and downloading a new one fixed the problem.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>