Fix Debian 8 image builds.
This applys a patch to the image's /etc/apt/sources-list to point to the
archived versions of the Debian 8 package repositories, now that the
ones the image used are unavailable.

A similar patch was done for Debian 9.

Testing:

Built Debian 8 DUE image with:
./due --create --platform linux/amd64 \
      --name example-debian-8 \
      --prompt ExD8 \
      --tag example-patch-debian-8 \
      --use-template example \
      --from debian:8 \
      --description 'Debian 8 with patch' \
      --image-patch debian/8/filesystem

Tested by checking out the 2021-08 branch of ONIE and building the
accton_as6701_32x target all and demo with:
 make MACHINEROOT=../machine/accton MACHINE=accton_as6701_32x all demo

Signed-off-by: Alex Doyle <adoyle@nvidia.com>