Example template README.md - collect image quirks.
A number of images have issues that need to be addressed before DUE
can work properly with them. The Example image seemed to be the best
place to put them since it's purpose is to just build an image with
DUE, and not specialize it to build packages, or anything else.

Currently:

Ubuntu 24.04
 With Ubuntu 24.04 introducing an 'ubuntu' user at ID 1000, the
 default add user behavior breaks ( in that the host system probably
 already _has_ a user 1000 who is, say, running DUE.)

 Since the best resolution for this is not obvious (does the account
 have special configuration?) , DUE throws an error and exits.

 From that point the user has a few options ( delete the user, steal it's
 ID, or log in as that user), all of which are documented in
 the Example README.md.

Old Debian images.
 Both Debian 8 and 9 reference package repositories that have been archived,
 so a patch is necessary to update their repository configuration to get them
 to build.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>