Add --delete-conflicting-user to --run
The Ubuntu 24 images have a user at 1000, which is the
most likely UID to create conflicts...and it does.

To get around this, the --run option now passes
--delete-conflicting-user to the container-create-user.sh
script that creates a user when the script is run.

This deletes the container's user and puts the invoking user
in its place.

If this causes conflicts, one can always log in as --user root
or specify a non conflicting user ID when running, or generate
an image that doesn't have the conflict.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>